app/requirements.txt
Click==7.0
Flask==1.1.1
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.4
PyMySQL==0.9.3
PyYAML==5.1.2
Werkzeug==0.16.0
